WOP-12297 EJ

Approved-by: Jyotsna Ramgoolam
Approved-by: kunal uttarwar
Approved-by: Jean-paulm
Approved-by: MuhammedArif

+++ app/src/main/java/za/co/woolworths/financial/services/android/models/dto/cart/DeliveryDetails.kt
@@ -2,6 +2,6 @@ package za.co.woolworths.financial.services.android.models.dto.cart
 
 class DeliveryDetails {
     var shippingAddress: ShippingAddress? =null
-    var shippingAmount: Int = 0
+    var shippingAmount: Double = 0.0
     var deliveryInfos: Array<DeliveryInfo>? =null
 }
